+++ lib/src/hash_algorithms.dart
@@ -1,4 +1,23 @@
-class HashAlgorithms {
+/// This class provides a set of constants, each representing a different hash algorithm.
+///
+/// * [SHA512] Represents the SHA-512 hash algorithm.
+/// * [BLAKE2B] Represents the Blake2b hash algorithm.
+/// * [MD2] Represents the MD2 hash algorithm.
+/// * [MD4] Represents the MD4 hash algorithm.
+/// * [MD5] Represents the MD5 hash algorithm.
+/// * [RIPEMD128] Represents the RIPEMD-128 hash algorithm.
+/// * [RIPEMD160] Represents the RIPEMD-160 hash algorithm.
+/// * [RIPEMD256] Represents the RIPEMD-256 hash algorithm.
+/// * [RIPEMD320] Represents the RIPEMD-320 hash algorithm.
+/// * [SHA1] Represents the SHA-1 hash algorithm.
+/// * [SHA3] Represents the SHA-3 hash algorithm.
+/// * [SHA224] Represents the SHA-224 hash algorithm.
+/// * [SHA256] Represents the SHA-256 hash algorithm.
+/// * [SHA384] Represents the SHA-384 hash algorithm.
+/// * [SM3] Represents the SM3 hash algorithm.
+/// * [TIGER] Represents the Tiger hash algorithm.
+/// * [WHIRLPOOL] Represents the Whirlpool hash algorithm.
+sealed class HashAlgorithms {
   static const String SHA512 = 'SHA-512';
   static const String BLAKE2B = 'Blake2b';
   static const String MD2 = 'MD2';