+++ lib/src/hashers/sha3.dart
@@ -2,7 +2,7 @@ import 'dart:typed_data';
 
 import 'package:crypto_dart/crypto_dart.dart';
 import 'package:crypto_dart/hashers.dart';
-import 'package:crypto_dart/src/encoders.dart';
+import 'package:crypto_dart/src/encoders/utf8.dart';
 import 'package:pointycastle/export.dart';
 
 class SHA3 extends Hasher {
@@ -17,7 +17,7 @@ class SHA3 extends Hasher {
   @override
   void init(dynamic data) {
     if (data is String) {
-      data = Encoders().UTF8.parse(data);
+      data = UTF8().parse(data);
     } else if (data is List<int>) {
       data = Uint8List.fromList(data);
     }
@@ -34,4 +34,3 @@ class SHA3 extends Hasher {
   @override
   String get algo => HashAlgorithms.SHA3;
 }
-