+++ lib/src/mode.dart
@@ -1,3 +1,4 @@
-// ignore_for_file: constant_identifier_names
-
+/// Represents the modes of operation for a block cipher.
+/// * [CBC] Cipher Block Chaining (CBC) mode.
+/// * [ECB] Electronic Codebook (ECB) mode
 enum Mode { CBC, ECB }