+++ lib/src/padding/padding.dart
@@ -1,5 +1,7 @@
-// ignore_for_file: constant_identifier_names
-
+/// Types of Padding
+/// * [PKCS7] - PKCS7 padding
+/// * [PKCS5] - PKCS5 padding
+/// * [ZEROPADDING] - Zero padding
 enum Padding {
   PKCS7,
   PKCS5,