+++ CHANGELOG.md
@@ -1,3 +1,15 @@
 ## 1.0.0
 
 - Initial version.
+
+## 1.0.3
+
+- made EvpKDF return EvpKDFResult rather than a list and also made ivSize optional for genrating key only.
+- PBKDF2 can now take blocklength from user (optional). thanks @indranildutta06  . 
+- add documentation.
+- add tests.
+- changed encoders names.
+- refactor some code.
+
+
+